Cleanup

+++ src/Vector/Vector.hpp
@@ -23,7 +23,7 @@ inline bool operator==(const Vector &r, const Vector &l)
     return (r.x == l.x && r.y == l.y);
 }
 
-inline double operator*(const Vector& r, const Vector& l)
+inline int64_t operator*(const Vector& r, const Vector& l)
 {
     return (r.x * l.x) + (r.y * l.y);
 }
